requirements.txt
Django==1.11.6
django-material==1.0.0
django-webpack-loader==0.5.0
django-phonenumber-field==1.3.0
django-db-geventpool==1.21
django-allauth==0.32.0
django-storages==1.6.4
sendgrid-django==4.0.4
googlemaps==2.5.0
google-cloud==0.26.1
pyzillow==0.5.5
sendgrid==3.6.3
usaddress==0.5.10
pytz==2017.2
pillow==4.2.1
gunicorn==19.7.1
gevent==1.2.2
geopy==1.11.0